+++ spec/spec_helper.rb
@@ -1,27 +1,39 @@
 # frozen_string_literal: true
 
 require 'bundler/setup'
-require 'json'
+require 'legion/logging'
+require 'legion/settings'
+require 'legion/cache/helper'
+require 'legion/crypt/helper'
+require 'legion/data/helper'
+require 'legion/json/helper'
+require 'legion/transport'
 
-unless defined?(Legion::Logging)
-  module Legion
-    module Logging
-      def self.debug(_msg); end
-
-      def self.info(_msg); end
-
-      def self.warn(_msg); end
-
-      def self.error(_msg); end
+module Legion
+  module Extensions
+    module Helpers
+      module Lex
+        include Legion::Logging::Helper
+        include Legion::Settings::Helper
+        include Legion::Cache::Helper
+        include Legion::Crypt::Helper
+        include Legion::Data::Helper
+        include Legion::JSON::Helper
+        include Legion::Transport::Helper
+      end
     end
-  end
-end
 
-unless defined?(Legion::Settings)
-  module Legion
-    module Settings
-      def self.[](_key)
-        {}
+    module Actors
+      class Every
+        include Helpers::Lex
+      end
+
+      class Once
+        include Helpers::Lex
+      end
+
+      class Subscription
+        include Helpers::Lex
       end
     end
   end